refactor(RoomView): split useRoomNavigation into useJumpToMessage and openThread/openRoom services
Rename the hook to useJumpToMessage to match the CONTEXT.md Positional State row.
Move Thread and Room opening into plain async services with explicit deps.
Isolate route-param consumption in a non-exported helper hook so the
generation counter has a single owner. Behavior and public surface unchanged;
the three existing suites pass with only import renames.

NATIVE-1550
